feat: re-export serve types from public mellea.* namespace
Creates a stable public API for m serve types by moving user-facing
types from cli/serve/models to mellea/serve/models. Users can now
import ChatMessage and related types via:

  from mellea.serve import ChatMessage

Changes:
- Add mellea/serve/ module with ChatMessage, TextContent,
  ImageUrlContent, and MessageContent
- Update cli/serve/models.py to re-export from mellea.serve
- Update m serve examples to use new import path
- Update m-serve.md documentation
- Maintain backward compatibility (cli.serve.models still works)

The new public API provides clear separation
between user-facing types (mellea.serve) and internal CLI
implementation (cli/serve).
