+++ mellea/__init__.py
@@ -1,8 +1,16 @@
 """Mellea."""
 
+from . import serve
 from .backends import model_ids
 from .stdlib.components.genstub import generative
 from .stdlib.session import MelleaSession, start_session
 from .stdlib.start_backend import start_backend
 
-__all__ = ["MelleaSession", "generative", "model_ids", "start_backend", "start_session"]
+__all__ = [
+    "MelleaSession",
+    "generative",
+    "model_ids",
+    "serve",
+    "start_backend",
+    "start_session",
+]
