+++ mellea/serve/__init__.py
@@ -1,0 +1,9 @@
+"""Public API for m serve types.
+
+This module provides the types that users need when writing serve() functions
+for use with the `m serve` command.
+"""
+
+from .models import ChatMessage, ImageUrlContent, MessageContent, TextContent
+
+__all__ = ["ChatMessage", "ImageUrlContent", "MessageContent", "TextContent"]
